Fix Annotation->Solr field mappings
There were some problems with the existing config parsing. The updated logic
allows for multiple mappings to be defined for an annotation type and/or Solr
field. Also, the SolrDocument code was using setField instead of addField
effectively eliminating the possibility of multi-valued fields.

Now,

    solr.f.foo = spam.eggs

will create a mapping for the "eggs" feature of the "spam" annotation type to
the Solr field "foo".

    solr.f.foo = spam.*
    solr.f.foo = spam

will both map the text from the "spam" annotation to the Solr field "foo".