Fix field mappings for Solr
There was a bug with the field mappings. It was using <annotation>.<feature> as
the key in the map and therefor not matching any annotations during the loop
later on.

Also, changed `setField` to `addField` to allow for multi-valued Solr fields.

Added a unit test for the field mappings as well.